# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AIAS Platform API
  description: Enterprise-grade AI consultancy platform API
  version: 1.0.0
  contact:
    name: AIAS Platform Support
    email: support@aias-platform.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://aias-platform.com/api
    description: Production
  - url: http://localhost:3000/api
    description: Development
...
security:
  - BearerAuth: []
paths:
...
  /v1/agents:
    get:
      tags: [Agents]
      summary: List agents
      parameters:
        - name: tenant_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by tenant ID
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                type: object
                properties:
                  agents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Agents]
      summary: Create agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreate'
      responses:
        '201':
          description: Created agent
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent:
                    $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [chatbot, automation, analytics, custom]
        config:
          type: object
        enabled:
          type: boolean
        created_at:
          type: string
          format: date-time
    AgentCreate:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
        type:
          type: string
          enum: [chatbot, automation, analytics, custom]
        config:
          type: object
        enabled:
          type: boolean
          default: true
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...